json update

+++ src/main/resources/static/js/deleteObj.js
@@ -2,38 +2,56 @@ const form = document.getElementById('deleteId');
 const surNameInp = document.getElementById('surName');
 const bookNameInp = document.getElementById('name');
 
-form.addEventListener('submit', async (event) =>{
+form.addEventListener('submit', async (event) => {
     event.preventDefault();
 
-    if(surNameInp.value == null){
-        const bookName = bookNameInp.value;
-    }
-    else{
-        const surName = surNameInp.value;
-    }
-    try{
-        const [response, response0] = await Promise.all([
-            fetch(`/api/v2/deleteUser/${surName}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({surName})
-        }),
-        fetch(`/api/v1/deleteBook/${bookName}`,{
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({bookName})
-        })
-    ]);
+    const fetchPromises = [];
 
-        if (!response.ok && !response0.ok) {
-            throw new Error(`HTTP error! status: ${response.status} ${response0.status}`);
+    let bookName = '';
+    let surName = '';
+
+    if (bookNameInp != null) {
+        bookName = bookNameInp.value;
+    }
+    if (surNameInp != null) {
+        surName = surNameInp.value;
+    }
+
+    try {
+        if (surName) {
+            fetchPromises.push(
+                fetch(`/api/v2/deleteUser/${surName}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify({ surName })
+                })
+            );
         }
-        console.log('Response data:', await response.json(), await response0.json());
-    }catch(error){
+
+        if (bookName) {
+            fetchPromises.push(
+                fetch(`/api/v1/deleteBook/${bookName}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify({ bookName })
+                })
+            );
+        }
+
+        const responses = await Promise.all(fetchPromises);
+
+        responses.forEach(async (response) => {
+            if (!response.ok) {
+                throw new Error(`HTTP error! status: ${response.status}`);
+            }
+            console.log('Response data:', await response.json());
+        });
+
+    } catch (error) {
         console.error('Error fetching name', error);
     }
 });
